refac: edit payroll period - remove holiday selection dialog and place the selection inside the create and edit page with collapsible button - add update confirmation dialog
